moved comps + scripts to `general`
